requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
dj-database-url
stripe
Pillow
scikit-learn
pandas
numpy
whitenoise
python-decouple
gunicorn
joblib
google-generativeai
requests
